requirements.txt
groq
google-genai
opencv-python
pyserial
sounddevice
pynput
numpy
flask
requests
